+++ packages/cubejs-backend-native/test/jinja.test.ts
@@ -2,27 +2,28 @@ import fs from 'fs';
 import path from 'path';
 
 import * as native from '../js';
+import { JinjaEngine, newJinjaEngine } from '../js';
 
 const suite = native.isFallbackBuild() ? xdescribe : describe;
 
-function loadTemplateFile(fileName: string): void {
+function loadTemplateFile(engine: native.JinjaEngine, fileName: string): void {
   const content = fs.readFileSync(path.join(process.cwd(), 'test', 'templates', fileName), 'utf8');
 
-  native.loadTemplate(fileName, content);
+  engine.loadTemplate(fileName, content);
 }
 
-function testTemplateBySnapshot(templateName: string, ctx: unknown) {
+function testTemplateBySnapshot(engine: JinjaEngine, templateName: string, ctx: unknown) {
   test(`render ${templateName}`, async () => {
-    const actual = native.renderTemplate(templateName, ctx);
+    const actual = engine.renderTemplate(templateName, ctx);
 
     expect(actual).toMatchSnapshot(templateName);
   });
 }
 
-function testLoadBrokenTemplateBySnapshot(templateName: string) {
+function testLoadBrokenTemplateBySnapshot(engine: JinjaEngine, templateName: string) {
   test(`render ${templateName}`, async () => {
     try {
-      loadTemplateFile(templateName);
+      loadTemplateFile(engine, templateName);
 
       throw new Error(`Template ${templateName} should throw an error!`);
     } catch (e) {
@@ -32,21 +33,20 @@ function testLoadBrokenTemplateBySnapshot(templateName: string) {
 }
 
 suite('Jinja', () => {
-  beforeAll(async () => {
-    native.initJinjaEngine({
-      debugInfo: true
-    });
-    native.clearTemplates();
+  const jinjaEngine = native.newJinjaEngine({
+    debugInfo: true
+  });
 
-    loadTemplateFile('.utils.jinja');
-    loadTemplateFile('dump_context.yml.jinja');
+  beforeAll(async () => {
+    loadTemplateFile(jinjaEngine, '.utils.jinja');
+    loadTemplateFile(jinjaEngine, 'dump_context.yml.jinja');
 
     for (let i = 1; i < 9; i++) {
-      loadTemplateFile(`0${i}.yml.jinja`);
+      loadTemplateFile(jinjaEngine, `0${i}.yml.jinja`);
     }
   });
 
-  testTemplateBySnapshot('dump_context.yml.jinja', {
+  testTemplateBySnapshot(jinjaEngine, 'dump_context.yml.jinja', {
     bool_true: true,
     bool_false: false,
     string: 'test string',
@@ -60,9 +60,9 @@ suite('Jinja', () => {
       userId: 1,
     }
   });
-  testLoadBrokenTemplateBySnapshot('template_error.jinja');
+  testLoadBrokenTemplateBySnapshot(jinjaEngine, 'template_error.jinja');
 
   for (let i = 1; i < 9; i++) {
-    testTemplateBySnapshot(`0${i}.yml.jinja`, {});
+    testTemplateBySnapshot(jinjaEngine, `0${i}.yml.jinja`, {});
   }
 });